Remove CatcherArea abstraction
Also fixes catcher size being relative to aspect ratio.
